Handling large reads data with filtlong

+++ modules/waterisk_modules.nf
@@ -1,6 +1,6 @@
 
 /*
-List all barcodes from miniON output
+List all barcodes contain in the input directory from miniON output
 */
 process IDENTIFIED_SAMPLES {
     input:
@@ -21,7 +21,7 @@ process IDENTIFIED_SAMPLES {
 /*
 Merge all seprates fastq.gz for each barcodes file into one file
 */
-process GZIP_FASTQ {
+process MERGE_SEPARATE_FASTQ {
     input:
     path barcode_dir
 
@@ -33,27 +33,6 @@ process GZIP_FASTQ {
     barID = barcode_dir.getSimpleName()
     """
         while read -r line; do cat \${line}/*.fastq.gz > ${barID}.fastq.gz; done < $barcode_dir
-    """
-}
-
-/*
-Run only fastqc
-*/
-process FASTQC {
-    input:
-    val barID
-    path fastq
-
-    output: 
-    val barID
-    path fastq
-
-    when:
-    params.only_qc == true
-
-    script:
-    """
-        //TODO
     """
 }
 
@@ -72,14 +51,9 @@ process REMOVE_BARCODES {
     path "trimmed_${barID}.fastq.gz"
 
     script:
-    if ( params.remove_barcode == true)
-        """
-        porechop --i $fastq -o trimmed_${barID}.fastq.gz -t ${task.cpus}
-        """
-    else
-        """
-        cp ${fastq} trimmed_${barID}.fastq.gz
-        """
+    """
+    porechop --i $fastq -o trimmed_${barID}.fastq.gz -t ${task.cpus}
+    """
 }
 
 /*
@@ -138,25 +112,73 @@ process CLEAN_READS {
 }
 
 /*
+Count the number of base
+*/
+process COUNT_BP {
+    
+    input: 
+    val barID
+    path trimmed_fastq
+
+    output:
+    val barID, emit: barID
+    env nbp, emit: number_bp
+    path trimmed_fastq, emit: fastq
+
+    script:
+    """
+    gzip -dc $trimmed_fastq > unzip.fastq
+    nbp=\$(awk 'NR % 4 == 0' ORS="" unzip.fastq | wc -m)
+    rm unzip.fastq
+    """
+}
+
+/*
+Remove the worst reads until only 500 Mbp remain, useful for very large read sets. If the input read set is less than 500 Mbp, this setting will have no effect.
+*/
+process SAMPLE_FASTQ {
+    publishDir "${params.output_dir}trimmed_output/"
+
+    input:
+    val barID
+    val nbp
+    path query
+
+    output:
+    val barID, emit: barID
+    path "${barID}sampleTrimmed.fastq.gz", emit: trimmed_fastq
+
+    script:
+    """
+    if [ ${nbp} -gt 50000000]; then
+        filtlong --min_length ${params.read_min_length} --target_bases 500000000 $query | gzip > "${barID}sampleTrimmed.fastq.gz"
+    else
+        cp $query "${barID}sampleTrimmed.fastq.gz"
+    fi
+    """
+}
+
+/*
 Assembling genome and plasmid with hybracter
 Hybracter also compare putative plasmid with PLSDB using MASH (see plassember_summary.tsv)
-TODO if incomplete assembly, write log info
 */
 process ASSEMBLE_GENOME { 
     label 'process_high'
     publishDir "${params.output_dir}genome_assembly/"
-    
+
     input:
     val barID
     path fastq
     
     output:
     val barID, emit: id
-    path "${barID}_plassembler_summary.tsv", emit: plassember_summary, optional: true
+    
+    //For complete assembly
     path "${barID}_sample_per_contig_stats.tsv", emit: assembly_summary, optional: true
+    path "${barID}_plassembler_summary.tsv", emit: plassember_summary, optional: true
     path "${barID}_sample_chromosome.fasta", emit: assembly_chr_fasta, optional: true
     path "${barID}_sample_plasmid.fasta", emit: assembly_pls_fasta, optional: true
-   
+    //For incomplete assembly. No contig size above the -c (minimal chrm size).
     path "${barID}_sample.fasta", emit: sample_fasta, optional: true
 
     script:
@@ -164,45 +186,31 @@ process ASSEMBLE_GENOME {
         """
         hybracter long-single -l $fastq -t ${task.cpus} --skip_qc --min_length ${params.read_min_length} --flyeModel --nano-hq -c ${params.c_size}
 
-        mv hybracter_out/processing/plassembler/sample/plassembler_summary.tsv ${barID}_plassembler_summary.tsv
-        [ -f hybracter_out/FINAL_OUTPUT/complete/sample_per_contig_stats.tsv ] && mv hybracter_out/FINAL_OUTPUT/complete/sample_per_contig_stats.tsv ${barID}_sample_per_contig_stats.tsv
-        [ -f hybracter_out/FINAL_OUTPUT/complete/sample_chromosome.fasta ] && mv hybracter_out/FINAL_OUTPUT/complete/sample_chromosome.fasta ${barID}_sample_chromosome.fasta 
-        [ -f hybracter_out/FINAL_OUTPUT/complete/sample_plasmid.fasta ] && mv hybracter_out/FINAL_OUTPUT/complete/sample_plasmid.fasta ${barID}_sample_plasmid.fasta
+        [ ! -f hybracter_out/processing/plassembler/sample/plassembler_summary.tsv ] || mv hybracter_out/processing/plassembler/sample/plassembler_summary.tsv ${barID}_plassembler_summary.tsv
+        [ ! -f hybracter_out/FINAL_OUTPUT/complete/sample_per_contig_stats.tsv ] || mv hybracter_out/FINAL_OUTPUT/complete/sample_per_contig_stats.tsv ${barID}_sample_per_contig_stats.tsv
+        [ ! -f hybracter_out/FINAL_OUTPUT/complete/sample_chromosome.fasta ] || mv hybracter_out/FINAL_OUTPUT/complete/sample_chromosome.fasta ${barID}_sample_chromosome.fasta 
+        [ ! -f hybracter_out/FINAL_OUTPUT/complete/sample_plasmid.fasta ] || mv hybracter_out/FINAL_OUTPUT/complete/sample_plasmid.fasta ${barID}_sample_plasmid.fasta
 
-        [ -f hybracter_out/FINAL_OUTPUT/incomplete/sample_per_contig_stats.tsv ] && mv hybracter_out/FINAL_OUTPUT/incomplete/sample_per_contig_stats.tsv ${barID}_sample_per_contig_stats.tsv
-        [ -f hybracter_out/FINAL_OUTPUT/incomplete/sample_final.fasta ] && mv hybracter_out/FINAL_OUTPUT/incomplete/sample_final.fasta ${barID}_sample.fasta
+        [ ! -f hybracter_out/FINAL_OUTPUT/incomplete/sample_final.fasta ] || mv hybracter_out/FINAL_OUTPUT/incomplete/sample_final.fasta ${barID}_sample.fasta
         """
     if (params.medaka == false)
         """
         hybracter long-single -l $fastq -t ${task.cpus} --skip_qc --no_medaka --min_length ${params.read_min_length} --flyeModel --nano-hq -c ${params.c_size}
-        mv hybracter_out/processing/plassembler/sample/plassembler_summary.tsv ${barID}_plassembler_summary.tsv
-        [ -f hybracter_out/FINAL_OUTPUT/complete/sample_per_contig_stats.tsv ] && mv hybracter_out/FINAL_OUTPUT/complete/sample_per_contig_stats.tsv ${barID}_sample_per_contig_stats.tsv
-        [ -f hybracter_out/FINAL_OUTPUT/complete/sample_chromosome.fasta ] && mv hybracter_out/FINAL_OUTPUT/complete/sample_chromosome.fasta ${barID}_sample_chromosome.fasta 
-        [ -f hybracter_out/FINAL_OUTPUT/complete/sample_plasmid.fasta ] && mv hybracter_out/FINAL_OUTPUT/complete/sample_plasmid.fasta ${barID}_sample_plasmid.fasta
+        
+        [ ! -f hybracter_out/processing/plassembler/sample/plassembler_summary.tsv ] || mv hybracter_out/processing/plassembler/sample/plassembler_summary.tsv ${barID}_plassembler_summary.tsv
+        [ ! -f hybracter_out/FINAL_OUTPUT/complete/sample_per_contig_stats.tsv ] || mv hybracter_out/FINAL_OUTPUT/complete/sample_per_contig_stats.tsv ${barID}_sample_per_contig_stats.tsv
+        [ ! -f hybracter_out/FINAL_OUTPUT/complete/sample_chromosome.fasta ] || mv hybracter_out/FINAL_OUTPUT/complete/sample_chromosome.fasta ${barID}_sample_chromosome.fasta 
+        [ ! -f hybracter_out/FINAL_OUTPUT/complete/sample_plasmid.fasta ] || mv hybracter_out/FINAL_OUTPUT/complete/sample_plasmid.fasta ${barID}_sample_plasmid.fasta
 
-        [ -f hybracter_out/FINAL_OUTPUT/incomplete/sample_per_contig_stats.tsv ] && mv hybracter_out/FINAL_OUTPUT/incomplete/sample_per_contig_stats.tsv ${barID}_sample_per_contig_stats.tsv
-        [ -f hybracter_out/FINAL_OUTPUT/incomplete/sample_final.fasta ] && mv hybracter_out/FINAL_OUTPUT/incomplete/sample_final.fasta ${barID}_sample.fasta
+        [ ! -f hybracter_out/FINAL_OUTPUT/incomplete/sample_final.fasta ] || && mv hybracter_out/FINAL_OUTPUT/incomplete/sample_final.fasta ${barID}_sample.fasta
         """
 
-    stub:
-    """
-    mkdir hybracter_out
-    mkdir hybracter_out/FINAL_OUTPUT
-    mkdir hybracter_out/FINAL_OUTPUT/complete
-    touch hybracter_out/FINAL_OUTPUT/complete/sample_summary.tsv
-    touch hybracter_out/FINAL_OUTPUT/complete/sample_chromosome.fasta
-    touch hybracter_out/FINAL_OUTPUT/complete/sample_plasmid.fasta
-    
-    mv hybracter_out/FINAL_OUTPUT/complete/sample_summary.tsv ${barID}_sample_per_contig_stats.tsv
-    mv hybracter_out/FINAL_OUTPUT/complete/sample_chromosome.fasta ${barID}_sample_chromosome.fasta 
-    mv hybracter_out/FINAL_OUTPUT/complete/sample_plasmid.fasta ${barID}_sample_plasmid.fasta
-    """
 }
 
 /*
 Identify AMR gene on plasmid and chromosome using abricate
 */
-process IDENTIFY_AMR_PLASMID {
+process IDENTIFY_AMR_PLASMID_COMPLETE {
     publishDir "${params.output_dir}plasmid_amr/"
 
     input:
@@ -218,7 +226,7 @@ process IDENTIFY_AMR_PLASMID {
     """
 }
 
-process IDENTIFY_AMR_CHRM {
+process IDENTIFY_AMR_CHRM_COMPLETE {
     publishDir "${params.output_dir}chrm_amr/"
 
     input:
